+++ instana/log.py
@@ -1,6 +1,5 @@
 from __future__ import print_function
 import os
-import sys
 import logging
 
 logger = None
@@ -32,53 +31,27 @@ def get_aws_lambda_logger():
     aws_lambda_logger.setLevel(logging.INFO)
     return aws_lambda_logger
 
-
-def running_in_gunicorn():
-    """
-    Determines if we are running inside of a gunicorn process and that the gunicorn logging package
-    is available.
-
-    @return:  Boolean
-    """
-    process_check = False
+def glogging_available():
     package_check = False
 
+    # Is the glogging package available?
     try:
-        # Is this a gunicorn process?
-        if hasattr(sys, 'argv'):
-            for arg in sys.argv:
-                if arg.find('gunicorn') >= 0:
-                    process_check = True
-        elif os.path.isfile("/proc/self/cmdline"):
-            with open("/proc/self/cmdline") as cmd:
-                contents = cmd.read()
+        from gunicorn import glogging
+    except ImportError:
+        pass
+    else:
+        package_check = True
+    
+    return package_check
 
-            parts = contents.split('\0')
-            parts.pop()
-            cmdline = " ".join(parts)
-
-            if cmdline.find('gunicorn') >= 0:
-                process_check = True
-
-        # Is the glogging package available?
-        try:
-            from gunicorn import glogging
-        except ImportError:
-            pass
-        else:
-            package_check = True
-
-        # Both have to be true for gunicorn logging
-        return process_check and package_check
-    except Exception as e:
-        print("Instana.log.running_in_gunicorn: %s", e, file=sys.stderr)
-        return False
-
+def gunicorn_logging_available():
+    from .util import running_in_gunicorn
+    return running_in_gunicorn() and glogging_available()
 
 aws_env = os.environ.get("AWS_EXECUTION_ENV", "")
 env_is_aws_lambda = "AWS_Lambda_" in aws_env
 
-if running_in_gunicorn():
+if gunicorn_logging_available():
     logger = logging.getLogger("gunicorn.error")
 elif env_is_aws_lambda is True:
     logger = get_aws_lambda_logger()
